shopfloor_base: fix access rights for public user
A SF app is responsible to handle its own authentication.
As such, a public user should be able to load the app using its url.
In v18 there's a mandatory check on READ access for all records
that are found by a model converter of a routing rule.

We don't want to expose to public all apps by default
and we don't want to add record rules to handle restrictions.
Our controllers should be allowed to publish a SF app.

Requires this fix in odoo core https://github.com/odoo/odoo/pull/221005